# PERF: special case numpy.dtype in is_extension_array_dtype