@jorisvandenbossche 2021-02-10
> assume you found this from existing asvs, if not can you add the one from above.

Yes, it came from profiling the concat benchmarks (https://github.com/pandas-dev/pandas/pull/39612), gives a small improvement. Now, apparently we don't have specific benchmarks for our dtype checks, so added one for `is_extension_array_dtype` (that could be supplemented with benchmarks for some of the other `is_` checkers, but going to leave that for another issue/PR).

> also a whatsnew mention is prob worth it.

Not sure it is worth it, this only gives a tiny improvement anywhere it is used (it was just lowing hanging fruit to shave off a percent, and with ArrayManager it gets used more, so getting more worth fixing it)